requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
streamlit-nightly
torch==1.6.0+cpu
torchvision==0.7.0+cpu
tqdm==4.43.0
transformers @ git+git://github.com/huggingface/transformers@59b5953d89544a66d73
